==> 1 - Creating a Model.ipynb <==
# Model Version Control with OCI Data Science

This project showcases how models can be versioned with the ADS SDK using OCI Data Science. Model Version Sets allow data science teams to continually improve deployed models and keep track of the version. Whilst this can be done using the OCI Console, practitioners may prefer to automate this using the SDK.

In this set of examples we'll showcase how you can:
- Register a model into the OCI Data Science Model Catalog
- Create a Version Set
- Add a new model to the Model Version Set
- Deploy a serverless batch jobs that loads the latest model version
- Delete and edit Model Version Sets

### Load Dependencies

import ads
import tempfile
from ads.model import SklearnModel
from ads.model import ModelVersionSet
from sklearn.datasets import load_iris
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from ads.common.model_metadata import UseCaseType

### Set OCIDs

project_ocid = 'ocid1.datascienceproject...'
compartment_ocid = 'ocid1.compartment...'

### Create Model Version Set

# Create a model version set
mvs = ModelVersionSet(
    name = "demo-model-version-set",
    description = "An example project showcasing model version set")
mvs.with_compartment_id(compartment_ocid).with_project_id(project_ocid).create()

model_version_set_id = 'ocid1.datasciencemodelversionset...'

### Create Basic Model

# Create a Sklearn model
iris = load_iris()
X, y = iris.data, iris.target
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25)
sklearn_estimator = LogisticRegression()
sklearn_estimator.fit(X_train, y_train)

### Create ADS Model Object

# Create an SklearnModel object
sklearn_model = SklearnModel(estimator=sklearn_estimator, artifact_dir='original-model/')


sklearn_model.prepare(inference_conda_env="generalml_p311_cpu_x86_64_v1",
training_conda_env="generalml_p311_cpu_x86_64_v1",
X_sample=X_train,
y_sample=y_train,
use_case_type=UseCaseType.MULTINOMIAL_CLASSIFICATION
)

sklearn_model.verify(X_test[:10])["prediction"]

sklearn_model.summary_status()

### Save Model to Catalog with Version Set

# Save the model and add it to the model version set
model_id = sklearn_model.save(compartment_id=compartment_ocid,
    project_id=project_ocid,
    display_name="Initial model",
    model_version_set=mvs,
    version_label="Version 1")

### View Models within Version Set

for item in ModelVersionSet.list(compartment_id=compartment_ocid,project_id=project_ocid):
    print(item)
    print("---------")

### Update Model Version Set Description

mvs.description = "Updated description of the model version set"
mvs.update()

### Save Train and Test Data to keep partitions consistent

import pandas as pd
pd.DataFrame(X_train).to_csv('data/x_train.csv',index=False)
pd.DataFrame(y_train).to_csv('data/y_train.csv',index=False)
pd.DataFrame(X_test).to_csv('data/x_test.csv',index=False)
pd.DataFrame(y_test).to_csv('data/y_test.csv',index=False)
