==> 2 - Updating a Model.ipynb <==
# Adding a new model to a Model Version Set

In this notebook we continue the example by:
- creating a new model
- uploading that model
- updating the model version set

This shows how, in a production scenario, we can continuously update the model artefact

### Load Dependencies

import ads
import tempfile
from ads.model import SklearnModel
from ads.model import ModelVersionSet
from sklearn.datasets import load_iris
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
import pandas as pd
import numpy as np
from ads.common.model_metadata import UseCaseType

### Load Current Champion Model

# load the model version set
mvs = ModelVersionSet.from_id(id='ocid1.datasciencemodelversionset...')

# view models in version set
models = mvs.models()

models

# get model ocid of entry in model version set
model_ocid = models[0].id

# download original model from model catalog
downloaded_model = SklearnModel.from_model_catalog(model_ocid,artifact_dir='downloaded-model/',ignore_conda_error=True)

### Calculate Goodness of Fit

import joblib

original_model = joblib.load('downloaded-model/model.joblib')

# load original train/test data
X_train = pd.read_csv('data/x_train.csv')
X_test = pd.read_csv('data/x_test.csv')
y_train = pd.read_csv('data/y_train.csv')
y_test = pd.read_csv('data/y_test.csv')

# score model on test data
preds = original_model.predict(X_test)

act_pred = pd.DataFrame({'actual':np.array(y_test['0']),'pred':preds})

print(classification_report(np.array(y_test['0']),preds))

Note: Change dataset later to something less deterministic

### Create New and Improved Model

rfc = RandomForestClassifier()

rfc = rfc.fit(X_train,y_train)

# check model performance

preds = rfc.predict(X_test)
print(classification_report(preds,y_test['0']))

### Save Model to Model Catalog with Version Set

# Create an SklearnModel object
sklearn_model = SklearnModel(estimator=rfc, artifact_dir='updated-model/')

sklearn_model.prepare(inference_conda_env="generalml_p311_cpu_x86_64_v1",
training_conda_env="generalml_p311_cpu_x86_64_v1",
X_sample=X_train,
y_sample=y_train,
use_case_type=UseCaseType.MULTINOMIAL_CLASSIFICATION
)

sklearn_model.verify(X_test[:10])["prediction"]

sklearn_model.summary_status()

# Save the model and add it to the model version set
project_ocid = 'ocid1.datascienceproject...'
compartment_ocid = 'ocid1.compartment...'

model_id = sklearn_model.save(compartment_id=compartment_ocid,
    project_id=project_ocid,
    display_name="Updated model",
    model_version_set=mvs,
    version_label="Version 2")

mvs.description = "Updated champion model to Random Forest"
mvs.update()
